refactor actions into separate files by type
