+++ src/Components/Nav.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 
-import { logout } from '../stateManagement/actions';
+import { logout } from '../stateManagement/actions/loginActions';
 
 const Nav = ({ username, isCompany, logout }) => {
   return (
